[MOVE] UI관련 데이터 패키지 이동
