+++ presentation/src/main/java/com/preonboarding/presentation/view/review/ReviewBindingAdapter.kt
@@ -9,7 +9,6 @@ import androidx.databinding.BindingAdapter
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.CenterCrop
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import com.preonboarding.domain.model.EditState
 
 @BindingAdapter("bindUri")
 fun ImageView.bindUri(uri: String?) {
