+++ presentation/src/main/java/com/preonboarding/presentation/view/review/ReviewUiState.kt
@@ -1,4 +1,4 @@
-package com.preonboarding.domain.model
+package com.preonboarding.presentation.view.review
 
 sealed class ReviewUiState {
